cleanup

+++ src/modules/jack/module-jack.h
@@ -62,7 +62,7 @@ enum {
 	SINK_MESSAGE_ON_SHUTDOWN
 };
 
-void* init_card(void* arg, const char *name, bool is_sink, unsigned channels);
+void* init_card(void* arg, const char *name, bool is_sink, uint8_t channels);
 void unload_card(void* arg,bool forced);
 static void timeout_cb(pa_mainloop_api*a, pa_time_event* e, const struct timeval *t, void *userdata);
 const char* get_merge_ref(pa_proplist *p, struct sBase *base);
